🚸 Minor touch calibration improvement (#26445)

+++ Marlin/src/lcd/marlinui.cpp
@@ -437,7 +437,7 @@ void MarlinUI::init() {
 
     #if HAS_TOUCH_BUTTONS
       uint8_t MarlinUI::touch_buttons;
-      uint8_t MarlinUI::repeat_delay;
+      uint16_t MarlinUI::repeat_delay;
     #endif
 
     #if EITHER(AUTO_BED_LEVELING_UBL, G26_MESH_VALIDATION)
